% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-conversion.r
\name{as.generic_mspct}
\alias{as.generic_mspct}
\alias{as.generic_mspct.colorSpec}
\title{Convert into generic_mspct}
\usage{
\method{as.generic_mspct}{colorSpec}(x, multiplier = 1, ...)
}
\arguments{
\item{x}{R object}

\item{multiplier}{numeric A multiplier to be applied to the spectral quantity
data to do unit or scale conversion.}

\item{...}{currently ignored.}
}
\description{
Convert into generic_mspct
}
