% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-conversion.r
\name{as.generic_spct}
\alias{as.generic_spct}
\alias{as.generic_spct.colorSpec}
\title{Coerce into generic_spct}
\usage{
\method{as.generic_spct}{colorSpec}(x, multiplier = 1, ...)
}
\arguments{
\item{x}{R object}

\item{multiplier}{numeric A multiplier to be applied to the spectral quantity
data to do unit or scale conversion.}

\item{...}{currently ignored.}
}
\description{
Coerce into generic_spct
}
