% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-asd-txt.r
\name{read_asd_tsv}
\alias{read_asd_tsv}
\title{Read File Saved by ASD's conversion tool}
\usage{
read_asd_tsv(
  file,
  date = NULL,
  geocode = NULL,
  label = NULL,
  tz = NULL,
  locale = readr::default_locale(),
  s.qty = NULL,
  scale.factor = 1,
  range = NULL
)
}
\arguments{
\item{file}{character string Path to the file to be read, following R's use
of forward slashes as separator for folder names.}

\item{date}{a \code{POSIXct} object to use to set the \code{"when.measured"}
attribute. If \code{NULL}, the default, the date is extracted from the file
header.}

\item{geocode}{A data frame with columns \code{lon} and \code{lat}, and
optionally \code{address} used to set attribute \code{"where.measured"}.}

\item{label}{character string to which to set the \code{"what.measured"}
attribute. If \code{NULL} the value of \code{basename(file)} is used, and
if \code{NA} the \code{"what.measured"} attribute is not set.}

\item{tz}{character A time zone recognized by R. If \code{NULL} it is
extracted from \code{locale}. If \code{""}, the default, the local time
zone is used.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), with time in UTC but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names. Its value must match that used to write the imported file, which is
not necessarily the default one or the local one.}

\item{s.qty}{character string The name of the quantity to be read as by
object constructors from 'photobiology'. If \code{NULL} quantity is guessed
from file header.}

\item{scale.factor}{numeric Multiplier to be applied to the spectral data so
that it matches \code{s.qty} in base of expression and units. In most cases
the default of 1.0 is correct.}

\item{range}{a numeric vector of length two, or any other object for which
function \code{range()} will return a range of wavelengths expressed in
nanometres.}
}
\value{
A \code{source_spct} object with a column \code{s.e.irrad} with
  spectral energy irradiance in \eqn{W m^{-2} nm^{-1}}, or a
  \code{reflector_spct} object with a column \code{Rfr} with fractional
  reflectance, or a \code{filter_spct} object with a column \code{Tfr} with
  fractional total transmittance. In all cases a column \code{w.length}
  contains wavelengths in nanometres and the attributes \code{comment},
  \code{what.measured}, \code{when.measured}, \code{how.measured},
  \code{where.measured} and \code{file.header} containing metadata for the
  spectrum.
}
\description{
Read the spectral data and parse the header of a energy irradiance, a
reflectance, a transmittance data, a raw-detector-counts data or arbitrary
data file as output by the text conversion tool for ASD spectrometers when
set for tab-, comma- or semicolon-separated output. ASD's field spectrometers
measure VIS and SWIR radiation.
}
\details{
The header of the file is first decoded and parsed to extract the
  time of data acquisition and serial number of the spectrometer, to locate
  the start of the spectral data and to detect the column separator used. The
  time in the header is the local time with no time zone information. Thus
  with the default \code{tz = ""} the time is re-expressed as UTC with a
  correction. When file import takes place on a different time zone than the
  measurement the measurement time zone must be supplied by the user as an
  argument to parameter \code{tz} or as part of the \code{locale} after
  setting \code{tz = NULL}. The metadata fields in the header are located by
  text matching.

  Spectral irradiance is returned as an object of class
  \code{\link[photobiology]{source_spct}}, reflectance as
  \code{\link[photobiology]{reflector_spct}}, transmittance as
  \code{\link[photobiology]{filter_spct}}, raw detector counts as
  \code{\link[photobiology]{raw_spct}}, and arbitrary spectral values as
  \code{\link[photobiology]{generic_spct}} objects with metadata stored in
  attributes \code{when.measured}, \code{what.measured}, and
  \code{how.measured}. The spectral quantity is guessed from the header
  metadata by default, but an argument to \code{s.qty} can be passed to
  override this default.

  The value stored in the \code{how.measured} attribute includes the
  spectrometer serial number extracted from the file header. If an argument
  is passed to parameter \code{geocode}, its value is saved in attribute
  \code{where.measured} (currently the geocode is not extracted from the file
  header). The object's \code{comment} always gives a text that includes the
  file name, time of import, function name and the version of packages
  'photobiology' and 'photobiologyInOut' used. The file header in whole is
  copied to attribute \code{file.header}.
}
\note{
Some files with ASD spectral data contain \code{nul} ASCII characters
  in the header. These are ignored with a warning while the data are still
  read. Depending on the program used to view these data files, only the top
  three lines will be displayed, or the file will be detected as containing
  binary data and not displayed.
}
\examples{

# energy spectral irradiance file

 file.name <-
   system.file("extdata", "asd-e-irrad-sky.tsv", 
               package = "photobiologyInOut", 
               mustWork = TRUE)
                
 asd.source_spct <- 
   read_asd_tsv(file = file.name,
                tz = "Europe/Helsinki")
 
 class_spct(asd.source_spct)
 summary(asd.source_spct)
 getWhenMeasured(asd.source_spct)
 getWhatMeasured(asd.source_spct)
 getHowMeasured(asd.source_spct)
 cat(comment(asd.source_spct))
 e_irrad(asd.source_spct)
 q_irrad(asd.source_spct, w.band = c(400, 700), 
         scale.factor = 1e6)
 
 asd_clipped.source_spct <- 
   read_asd_tsv(file = file.name,
                tz = "Europe/Helsinki",
                range = c(400, 700))
 
 class_spct(asd_clipped.source_spct)
 summary(asd_clipped.source_spct)

# spectral reflectance file

 file.name <-
   system.file("extdata", "asd-Rfr-panel-50pc.tsv", 
               package = "photobiologyInOut",
               mustWork = TRUE)
                
 asd.reflector_spct <- 
   read_asd_tsv(file = file.name,
                tz = "Europe/Helsinki")
 
 class_spct(asd.reflector_spct)
 summary(asd.reflector_spct)
 getWhenMeasured(asd.reflector_spct)
 getWhatMeasured(asd.reflector_spct)
 getHowMeasured(asd.reflector_spct)
 cat(comment(asd.reflector_spct))

 asd_clipped.reflector_spct <- 
   read_asd_tsv(file = file.name,
               tz = "Europe/Helsinki",
               range = c(400, 700))
 
 class_spct(asd_clipped.reflector_spct)
 summary(asd_clipped.reflector_spct)
 
 # Raw-counts data

 file.name <-
   system.file("extdata", "asd-raw-gravel.tsv", 
               package = "photobiologyInOut",
               mustWork = TRUE)
                
 asd.raw_spct <- 
   read_asd_tsv(file = file.name,
                tz = "Europe/Helsinki")
 
 class_spct(asd.raw_spct)
 summary(asd.raw_spct)
 getWhenMeasured(asd.raw_spct)
 getWhatMeasured(asd.raw_spct)
 getHowMeasured(asd.raw_spct)
 cat(comment(asd.raw_spct))

}
\references{
\url{https://www.malvernpanalytical.com/}
}
