% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-asd-txt.r
\name{read_asdtxt}
\alias{read_asdtxt}
\title{Read File Saved by ASD's conversion tool}
\usage{
read_asdtxt(
  file,
  qty.in = NULL,
  scale.factor = 1,
  date = NULL,
  geocode = NULL,
  label = NULL,
  tz = NULL,
  locale = readr::default_locale(),
  range = NULL
)
}
\arguments{
\item{file}{character string Path to the file to be read, following R's use
of forward slashes as separator for folder names.}

\item{qty.in}{character string The name of the quantity to be read as by
object constructors from 'photobiology'. If \code{NULL} quantity is guessed
from file header.}

\item{scale.factor}{numeric Multiplier to be applied to the spectral data so
that it matches \code{qty.in} in base of expression and units. In most cases
the default of 1.0 is correct.}

\item{date}{a \code{POSIXct} object to use to set the \code{"when.measured"}
attribute. If \code{NULL}, the default, the date is extracted from the file
header.}

\item{geocode}{A data frame with columns \code{lon} and \code{lat}, and
optionally \code{address} used to set attribute \code{"where.measured"}.}

\item{label}{character string to which to set the \code{"what.measured"}
attribute. If \code{NULL} the value of \code{basename(file)} is used, and
if \code{NA} the \code{"what.measured"} attribute is not set.}

\item{tz}{character A time zone recognized by R. If \code{NULL}, the default,
it is extracted from `locale`.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names. Its value must match that used to write the imported file, which is
not necessarily the default one or the local one.}

\item{range}{a numeric vector of length two, or any other object for which
function \code{range()} will return a range of wavelengths expressed in
nanometres.}
}
\value{
A \code{source_spct} object with a column \code{s.e.irrad} with
  spectral energy irradiance in \eqn{W m^{-2} nm^{-1}}, or a
  \code{reflector_spct} object with a column \code{Rfr} with fractional
  reflectance, or a \code{filter_spct} object with a column \code{Tfr} with
  fractional total transmittance. In all cases a column \code{w.length}
  contains wavelengths in nanometres and the attributes \code{comment},
  \code{what.measured}, \code{when.measured}, \code{how.measured},
  \code{where.measured} and \code{file.header} containing metadata for the
  spectrum.
}
\description{
Read the spectral data and parse the header of a energy irradiance, a
reflectance or a transmittance data file as output by the text conversion
tool for ASD spectrometers. ASD's field spectrometers measure VIS and SWIR
radiation. The spectral irradiance text files seem to be tab plus space
delimited.
}
\details{
The header of the file is first decoded and parsed to extract the
  time of data acquisition and serial number of the spectrometer, and to
  locate the start of the spectral data. The fields are located by name.
  The spectral irradiance is re-expressed in \eqn{W m^{-2} nm^{-1}} and 
  returned as an object of class \code{\link[photobiology]{source_spct}} 
  with metadata stored in attributes \code{when.measured},
  \code{what.measured}, and \code{how.measured} set to values extracted from
  the header. The value stored in the \code{how.measured} attribute includes
  the spectrometer serial number extracted from the file header. 
  If an argument is passed to parameter \code{geocode}, its value is
  saved in attribute \code{where.measured}. The file header in whole
  is copied to attribute \code{file.header}. The object's \code{comment}
  always gives a text that includes the file name, time of import, function
  name and the version of packages 'photobiology' and 'photobiologyInOut' 
  used.
}
\examples{

# energy spectral irradiance file

 file.name <-
   system.file("extdata", "irrad-sky.asd.txt", 
               package = "photobiologyInOut", mustWork = TRUE)
                
 asd.source_spct <- 
   read_asdtxt(file = file.name,
                   locale = readr::locale("en", 
                                          decimal_mark = ",",
                                          grouping_mark = "",
                                          tz = "Europe/Helsinki"))
 
 class_spct(asd.source_spct)
 summary(asd.source_spct)
 getWhenMeasured(asd.source_spct)
 getWhatMeasured(asd.source_spct)
 getHowMeasured(asd.source_spct)
 cat(comment(asd.source_spct))
 e_irrad(asd.source_spct)
 q_irrad(asd.source_spct, w.band = c(400, 700), 
         scale.factor = 1e6)
 
 asd_clipped.source_spct <- 
   read_asdtxt(file = file.name,
               locale = readr::locale("en", 
                                      decimal_mark = ",",
                                      grouping_mark = "",
                                      tz = "Europe/Helsinki"),
               range = c(400, 700))
 
 class_spct(asd_clipped.source_spct)
 summary(asd_clipped.source_spct)

# spectral reflectance file

 file.name <-
   system.file("extdata", "reflec-panel-50pc.asd.txt", 
               package = "photobiologyInOut", mustWork = TRUE)
                
 asd.reflector_spct <- 
   read_asdtxt(file = file.name,
               locale = readr::locale("en", 
                                      decimal_mark = ",",
                                      grouping_mark = "",
                                      tz = "Europe/Helsinki"))
 
 class_spct(asd.reflector_spct)
 summary(asd.reflector_spct)
 getWhenMeasured(asd.reflector_spct)
 getWhatMeasured(asd.reflector_spct)
 getHowMeasured(asd.reflector_spct)
 cat(comment(asd.reflector_spct))

 asd_clipped.reflector_spct <- 
   read_asdtxt(file = file.name,
               locale = readr::locale("en", 
                                      decimal_mark = ",",
                                      grouping_mark = "",
                                      tz = "Europe/Helsinki"),
               range = c(400, 700))
 
 class_spct(asd_clipped.reflector_spct)
 summary(asd_clipped.reflector_spct)

}
\references{
\url{https://www.malvernpanalytical.com/}
}
