% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-macam-dta.r
\name{read_macam_dta}
\alias{read_macam_dta}
\title{Read \code{.DTA} File Saved by Macam's Software.}
\usage{
read_macam_dta(file, date = NULL, geocode = NULL, label = NULL,
  tz = NULL, locale = readr::default_locale())
}
\arguments{
\item{file}{character string}

\item{date}{a \code{POSIXct} object, but if \code{NULL} the date stored in
file is used, and if \code{NA} no date variable is added}

\item{geocode}{A data frame with columns \code{lon} and \code{lat}.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{tz}{character Time zone used for interpreting times saved in the
file header.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\value{
A source_spct object.
}
\description{
Reads and parses the header of a processed data file as output by the PC
program to extract the time and date fields and a user label if present,
and then imports wavelengths and spectral energy irradiance values.
}
\author{
Pedro J. Aphalo
}
\references{
\url{http://www.r4photobiology.info}
}
\keyword{misc}

