% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-walz-lsa-xlsx.r
\name{read_walz_lsa_xlsx}
\alias{read_walz_lsa_xlsx}
\title{Read '.xlsx' file(s) saved by Walz's LSA-2050 leaf analyser}
\usage{
read_walz_lsa_xlsx(
  file,
  label = NULL,
  data_skip = 0,
  n_max = Inf,
  na = c("", "NA", "NAN"),
  marker.rename = NULL,
  returned.data = "consolidated",
  drop.positions = NULL,
  ...
)
}
\arguments{
\item{file}{Path to file as a character string.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{data_skip}{integer Number of records (rows) to skip from the actual
data block.}

\item{n_max}{integer Maximum number of records to read.}

\item{na}{character Vector of strings to interpret as missing values. Set 
this option to character() to indicate no missing values.}

\item{marker.rename}{character A named vector with new \code{Marker} values 
named according to the Markers set in the LSA-2050.}

\item{returned.data}{character If \code{"consolidated"} return all data in a
single data frame, with the "SAT chart" data as a list column containing
one data frame per sample. Other values return a data frame containing a
subset of the columns: with \code{"measured"}, return only the data in the
worksheet named "Measured", with \code{"visible"} return only data from
columns visible in the worksheet named "Measured", and \code{"raw"} returns
only data stored in both worksheets as numeric values, i.e., ignores values
computed by formulas included in the worksheet. \emph{For debugging:} if
\code{"list"}, return a named list with two member data frames, each
containing the data from one worksheet.}

\item{drop.positions}{logical If \code{TRUE} the columns related to leaf
position are deleted, and if \code{FALSE} they are retained. If \code{NULL}
they are dropped only if they contain no finite data.}

\item{...}{Further named arguments currently passed to 
\code{\link[readxl]{read_excel}()}.}
}
\value{
\code{read_csi_dat()} returns a \code{tibble::tibble} object.

The returned value depends on the argument passed to parameter
  \code{returned.data}. With the default, \code{"consolidated"}, all the data
  from both worksheets are returned in a single data frame with 44 columns.
  With arguments \code{"measured"} (43 columns), \code{"visible"} (17
  columns) and \code{"raw"} (24 columns) different variable subsets from the
  consolidated data frame are returned. Finally for debugging, with
  \code{"list"}, a list containing one data frame with the data from each
  worksheet is returned.
}
\description{
Read the two worksheets contained in the workbook file downloaded by Walz's
LSA software from an LSA-2050 instrument. Combine the data into a single
data frame and add metadata as its attributes.
}
\details{
The worksheets contain formulas for computations, on import the
  results of the computations are computed and stored in the returned object.
  The raw readings and calibration values are also imported as stored as an
  attribute. The names of the worksheets are preserved, and if the do not
  match the expected values, are warning is issued. Data manipulation is
  based on column names rather than positions. As the \code{Date} and
  \code{Time} columns contain identical data (formatted differently) for
  efficiency, only \code{Time} is retained in the returned object (after
  testing for equality). Units and bases of expression are not modified,
  but some column names are edited to comply with R naming rules.

  The named vector passed as argument to \code{marker.rename} if not
  \code{NULL}, is used to rename the values set as Marker by the LSA-2050
  (single capital letters) into informative character strings.
  \code{marker.rename} can contain a superset of the required mappings but a
  subset is not accepted.
}
\note{
This function has been tested with the output from LSA software
  version 1.09 and an example data file downloaded from Walz's website.
}
\section{Warning!}{
 The computed values for the formulas in the worksheets
  need to be present in the Excel workbook for import to succeed. If they are
  missing, zeros will appear in the worksheet in place of the results from
  calculations. The workbooks saved using the LSA software contain the
  computed cached values. Opening and re-saving the file can result in the
  loss of the cached computed values. Importing such a file triggers a
  message. The cached values from formulas can be restores in Excel or in
  LibreOffice. In LibreOffice open the workbook file, use \strong{data >
  calculate > recalculate hard} to recalculate all formulas and then save the
  file. (In LibreOffice a manually triggered hard recalculation is also
  needed for the correct computed values to be displayed.) The example file
  used in the examples and tests has been resaved in LibreOffice.

  To avoid depending on the import of cached values stored in the worksheets,
  one possible solution would be to reimplement all computations in R after
  importing the "raw" data (plus calibration parameters). These are stored in
  the worksheet as numeric values. This would provide a failure-proof
  approach, but updates to equations used in the worksheets would require
  matching updates to the R code. Passing \code{returned.data = "raw"} in the
  call imports only these values.
}

\examples{

 file.name <-
   system.file("extdata", "walz-lsa-2050-prunus.xlsx", 
               package = "photobiologyInOut", 
               mustWork = TRUE)
               
 lsa.df1 <- read_walz_lsa_xlsx(file.name)
 colnames(lsa.df1)
 plot(lsa.df1$SAT.F.ls[[1]], type = "b")
 cat(comment(lsa.df1))
 
 lsa.df2 <-
  read_walz_lsa_xlsx(file.name,
                     returned.data = "visible",
                     marker.rename = c(A = "adaxial", B = "abaxial"))
 colnames(lsa.df2)
 cat(comment(lsa.df2))
 
}
\references{
\url{https://www.campbellsci.eu/}
}
