% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-wasatch-csv.r
\name{read_wasatch_csv}
\alias{read_wasatch_csv}
\title{Read File Saved by Wasatch's Enlighten.}
\usage{
read_wasatch_csv(
  file,
  scale.factor = 1,
  date = NULL,
  geocode = NULL,
  label = NULL,
  tz = NULL,
  locale = readr::default_locale()
)
}
\arguments{
\item{file}{character}

\item{scale.factor}{numeric vector of length 1, or length equal to the number
of rows (= detector pixels). Numeric multiplier applied to returned
spectral values.}

\item{date}{a \code{POSIXct} object to use to set the \code{"when.measured"}
attribute. If \code{NULL}, the default, the date is extracted from the
file header.}

\item{geocode}{A data frame with columns \code{lon} and \code{lat} used to
set attribute \code{"where.measured"}.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{tz}{character Time zone is by default read from the file.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\value{
A source_spct object.
}
\description{
Reads and parses the header of a processed data file as output by
Enlighten to extract the whole header field. The time field is
retrieved and decoded as well additional metadata.
}
\note{
Tested only with two example files.
}
\references{
\url{https://wasatchphotonics.com/}
  \url{https://wasatchphotonics.com/product-category/software/}
}
\keyword{misc}
