% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-conversion.r
\name{rspec2mspct}
\alias{rspec2mspct}
\alias{rspec2spct}
\title{Convert "pavo::rspec" objects}
\usage{
rspec2mspct(
  x,
  member.class = "reflector_spct",
  spct.data.var = "Rpc",
  multiplier = 1,
  ...
)

rspec2spct(x, multiplier = 1, ...)
}
\arguments{
\item{x}{rspec object}

\item{member.class}{character One of the spectrum classes defined in package
'photobiology'. Examples are `reflector_spct`, `filter_spct`,
`source_spct`, etc.}

\item{spct.data.var}{character The name to be used for the 'spc' data when
constructing the spectral objects. Examples are spectral reflectance as
percent or as a fraction of one, `Rpc`, `Rfr`, spectral transmittance`Tpc`,
`Tfr`, spectral energy irradiance in \eqn{w m^{-2}} `s.e.irrad`, spectral
photon irradiance in \eqn{mol m^{-2} s^{-1}} `s.q.irrad`, etc.}

\item{multiplier}{numeric A multiplier to be applied to spectral data in the
'rspc' object to do unit or scale conversion.}

\item{...}{currently ignored.}
}
\description{
Convert between objects of class \code{rspec} as used in package 'pavo' and
similar objects returned by functions from package 'lightr' in objects of
classes from 'photobiology'
}
\details{
\code{rspec} objects contain data for multiple spectra in wide form,
  i.e., one column per spectrum. Function \code{rscpt2mspct()} converts them
  into a list-like collection of spectra (xxxx_mspct) with spectra stored as
  separate members. In contrast, \code{rscpt2spct()} stores the spectra in
  long form into a single data-frame-like object (xxxx_spct). The classes of
  the returned objects are defined in package 'photobiology' and derived from
  \code{list} and \code{data.frame} respectively.

  Objects of class \code{pavo::rspec} do not contain metadata or class 
  data from which the quantity measured and the units of expression could be 
  retrieved. Thus, when using this function users need to use parameter 
  \code{multiplier} to convert the data to the units and scaling that is 
  expected by the object constructors defined in package 'photobiology' and
  use parameter \code{spct.data.var} to indicate the physical quantity.
  
 @note \code{pavo::rspec} objects may use memory more efficiently than
   spectral objects of the classes for collections of spectra defined in
   package 'photobiology' as in 'pavo' wavelengths are assumed to be the same for all
   spectra stored in an object. So, wavelengths are always stored only once 
   per object. In contrast, 'photobiology' avoids this assumption allowing
   heterogeneous collections containing member spectra with different wavelengths 
   and lengths. This requires storing Wavelengths for each spectrum, but as
   spectral classes are derived from 'tbl_df' in many cases redundant
   copies of wavelength data are avoided by optimizations in 'tibble'.
}
\section{Warning!}{
 Always check the sanity of the imported or exported data 
  values, as guessing is needed when matching the different classes, and the 
  functions defined here are NOT guaranteed to return valid data without help
  from the user through optional function arguments.
}

\examples{
# example run only if 'pavo' is available
if (requireNamespace("pavo", quietly = TRUE)) {
  library(pavo)
  data(sicalis, package = "pavo")
  sicalis.mspct <- rspec2mspct(sicalis)
  class(sicalis.mspct)

  data(teal, package = "pavo")
  teal.spct <- rspec2spct(teal)
  class(teal.spct)
  levels(teal.spct[["spct.idx"]])
  angles <- seq(from = 15, to = 75, by = 5) # from teal's documentation
  teal.spct[["angle"]] <- angles[as.numeric(teal.spct[["spct.idx"]])]
  teal.spct
}

}
