% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-conversion.r
\name{mspct2fdata}
\alias{mspct2fdata}
\alias{spct2fdata}
\title{Convert spectra into 'fda.usc::fdata' objects}
\usage{
mspct2fdata(x, spct.data.var = NULL, multiplier = 1, ...)

spct2fdata(x, spct.data.var = NULL, multiplier = 1, ...)
}
\arguments{
\item{x}{generic_mspct or generic_spct object.}

\item{spct.data.var}{character The name of the column containing data to
export. If \code{NULL} the first spectral data column found is used.}

\item{multiplier}{numeric A multiplier to be applied to the 'spc' data to do
unit or scale conversion. For example "a.u." units in some examples in
package 'hyperSpec' seem to have scale factors applied.}

\item{...}{currently ignored.}
}
\description{
Convert spectral objects (xxxx_spct, xxxx_mspct) as defined in package
'photobiology' into \code{fda.usc::fdata} objects, preserving as much
information as possible. As \code{fdata} objects can contain many other kinds
of data, currently only export is supported.
}
\section{Warning!}{
 Always check the sanity of the imported or exported data
  values, as guessing is needed when matching the different classes, and the
  functions defined here are NOT guaranteed to return valid data whitout help
  from the user through optional function arguments.
}

\examples{
if (requireNamespace("fda.usc", quietly = TRUE)) {
  sun.fdata <- spct2fdata(sun.spct)
  str(sun.fdata)
  polyester.fdata <- spct2fdata(polyester.spct)
  str(polyester.fdata)
}

}
