% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobiologyInOut.r
\docType{package}
\name{photobiologyInOut-package}
\alias{photobiologyInOut}
\alias{photobiologyInOut-package}
\title{photobiologyInOut: Read Spectral and Logged Data from Foreign Files}
\description{
Functions for reading, and in some cases writing, foreign files
containing spectral data from spectrometers and their associated software,
output from daylight simulation models in common use, and some spectral
data repositories. As well as functions for exchange of spectral data with
other R packages.
}
\note{
From version 0.4.4 the time zone (tz) used for decoding dates
  and times in files imported defaults to "UTC". In most cases you will need
  to pass the tz (or the locale) where the file was created as an argument to
  the functions!
}
\section{Warning!}{
 Most of the file formats supported are not standardized,
  and are a moving target because of changes in instrument firmware and
  support software. In addition the output format, especially with models,
  can depend on settings that users can alter. So do check that import is
  working as expected, and if not, please please raise a bug or support
  ticket and upload one example of an incorrectly decoded file.
}

\seealso{
Useful links:
\itemize{
  \item \url{http://www.r4photobiology.info/}
  \item Report bugs at \url{https://bitbucket.org/aphalo/photobiologyinout/issues/}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi}

Other contributors:
\itemize{
  \item Titta K. Kotilainen \email{titta.kotilainen@helsinki.fi} [contributor]
}

}
