% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cid-spectravue-csv.r
\name{read_cid_spectravue_csv}
\alias{read_cid_spectravue_csv}
\title{Read File Saved by CID's SpectraVue.}
\usage{
read_cid_spectravue_csv(
  file,
  date = NULL,
  geocode = NULL,
  label = NULL,
  tz = NULL,
  locale = readr::default_locale(),
  scale.factor = 1,
  ...
)
}
\arguments{
\item{file}{character}

\item{date}{a \code{POSIXct} object to use to set the \code{"when.measured"}
attribute. If \code{NULL}, the default, the date and time are extracted
from the file.}

\item{geocode}{A data frame with columns \code{lon} and \code{lat} used to
set attribute \code{"where.measured"}. If \code{NULL}, the default, the
geocode is extracted from the file, if present, and if \code{NA} the
"where.measured" attribute is not set.}

\item{label}{character string. If \code{NULL}, the default, the value of the
"tag" present in the file is used, and if \code{NA} the "what.measured"
attribute is not set.}

\item{tz}{character Time zone is by default that of the machine's locale.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{scale.factor}{numeric vector of length 1, or length equal to the number
of data rows (= number of spectra). Numeric multiplier applied to returned
spectral values.}

\item{...}{additional arguments passed to the constructor of the `filter_spct`
object.}
}
\value{
An object of class \code{filter_spct}.
}
\description{
Read wavelength and spectral data from Measurements.CSV files exported from CID
Bio-Sciences' SpectraVue CI-710s leaf spectrometer, importing them into R.
Available metadata is also extracted from the file.
}
\details{
\code{read_cid_spectravue_csv()} only accepts "row oriented" CSV files. These
may contain multiple spectra, one per row.

SpectrVue's row-wise spectral Measurements.CSV files contain columns
  with metadata on the right edge, followed by columns with data for each
  wavelength. The value in column "Mode" indicates the quantity measured,
  decoded into Tfr, Rfr or A. The data in each row in the CSV file are read
  and stored in a `filter_spct` object. These objects are collected into a
  single `filter_mspct` object to be returned.
}
\note{
SpectraVue creates four files for each measurement, from these, this
  function reads the one with name ending in "Measurements.CSV". The first
  part of the file name gives the time of measurement.
}
\references{
\url{https://cid-inc.com/}
}
