% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ooov-file.r
\name{read_oo_ovirrad}
\alias{read_oo_ovirrad}
\title{Read File Saved by Ocean Optics' OceanView.}
\usage{
read_oo_ovirrad(
  file,
  date = NULL,
  geocode = NULL,
  label = NULL,
  tz = NULL,
  locale = readr::default_locale(),
  range = NULL
)
}
\arguments{
\item{file}{character string}

\item{date}{a \code{POSIXct} object to use to set the \code{"when.measured"}
attribute. If \code{NULL}, the default, the date is extracted from the
file header.}

\item{geocode}{A data frame with columns \code{lon} and \code{lat} used to
set attribute \code{"where.measured"}.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{tz}{character Time zone is by default read from the file.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{range}{a numeric vector of length two, or any other object for which 
function \code{range()} will return a range of wavelengths expressed in 
nanometres.}
}
\value{
A source_spct object.
}
\description{
Reads and parses the header of a processed data file as output by
OceanView to extract the whole header remark field. The time field is
retrieved and decoded. OceanView is a program from Ocean Optics that 
replaces the no longer supported SpectraSuite program.
}
\examples{

 file.name <- 
   system.file("extdata", "spectrum.OVIrrad", 
               package = "photobiologyInOut", mustWork = TRUE)
                
 ooov.spct <- 
   read_oo_ovirrad(file = file.name,
                   locale = readr::locale("en", 
                                          decimal_mark = ",",
                                          grouping_mark = "",
                                          tz = "Europe/Paris"))
 
 ooov.spct
 getWhenMeasured(ooov.spct)
 getWhatMeasured(ooov.spct)
 getHowMeasured(ooov.spct)
 cat(comment(ooov.spct))

 ooov_clipped.spct <- 
   read_oo_ovirrad(file = file.name,
                   locale = readr::locale("en", 
                                          decimal_mark = ",",
                                          grouping_mark = "",
                                          tz = "Europe/Paris"),
                   range = c(280, NA))
 
 ooov_clipped.spct
 getWhenMeasured(ooov_clipped.spct)
 getWhatMeasured(ooov_clipped.spct)
 getHowMeasured(ooov_clipped.spct)
 cat(comment(ooov_clipped.spct))

}
\references{
\url{https://www.oceanoptics.com/}
}
\keyword{misc}
